[CHORE]: Refactorización de componente de contribuciones
