--- frontend/src/repositories/repository/folder/sections/folderContributions.css
@@ -1,58 +1,0 @@
-/* App.css */
-.button {
-    padding: 10px 20px;
-    font-size: 16px;
-    color: white;
-    background-color: #007BFF;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-  }
-  
-  .button:disabled {
-    background-color: #6C757D;
-    cursor: not-allowed;
-  }
-  
-  .spinner {
-    display: inline-block;
-    width: 20px;
-    height: 20px;
-    border: 3px solid rgba(0, 0, 0, 0.1);
-    border-top: 3px solid #ffffff;
-    border-radius: 50%;
-    animation: spin 1s linear infinite;
-  }
-  
-  @keyframes spin {
-    from {
-      transform: rotate(0deg);
-    }
-    to {
-      transform: rotate(360deg);
-    }
-  }
-  
-  .medal {
-    font-size: 2rem;
-}
-
-.medal.gold {
-    color: #FFD700;
-}
-
-.medal.silver {
-    color: #C0C0C0;
-}
-
-.medal.bronze {
-    color: #CD7F32;
-}
-
-.medal.dotted {
-    color: transparent;
-    border: 2px dashed #999;
-    border-radius: 50%;
-    padding: 4px;
-}
-
